fix warning in mandatory

+++ mandatory/src/parsing/parse.c
@@ -19,8 +19,8 @@ enum e_side	get_direction(char c)
 
 static void	invert_map_rows(t_game *g)
 {
-	int	i;
-	int	j;
+	unsigned long	i;
+	unsigned long	j;
 
 	i = 0;
 	while (i < g->map.height / 2)
